tools/refengine/indextemplate: Move from tools/indextemplate
There's no technical reason to locate a ref-engine implementation
package under refengine/, but doing so should reduce clutter in tools/
and make it a bit easier for newcomers to understand.
