+++ tools/cmd/main.go
@@ -20,7 +20,7 @@ import (
 	"github.com/sirupsen/logrus"
 	"github.com/urfave/cli"
 
-	_ "github.com/xiekeyang/oci-discovery/tools/indextemplate"
+	_ "github.com/xiekeyang/oci-discovery/tools/refengine/indextemplate"
 )
 
 func main() {
